Extract view holder and remove unnecessary thumbnail optimisation
